Add files via upload
This Python program uses OpenCV to detect and count screws in an image using a combination of image processing techniques.
It starts by reading the input image and resizing it for consistent display.
 The image is converted to grayscale and enhanced using CLAHE (Contrast Limited Adaptive Histogram Equalization) to improve contrast.
 Adaptive thresholding is applied to convert the image to binary format, highlighting potential screw regions. 
Morphological operations (closing and opening) are then used to clean noise and connect fragmented parts. 
Contours are detected, and each is filtered based on area, solidity, and aspect ratio to eliminate irrelevanvt shapes. Valid contours, presumed to be screws, are drawn on the image, and their count is displayed. The final result, along with intermediate steps , is shown . The function returns the number of screws detected in the image.